Bug when using array of CIDR blocks
When an array of CIDR blocks were given, if the first block did not
contain the ip address then notBannedIp would be set to true.  Then when
the next loop through the CIDR blocks occured, even though the IP was
found and was on a blacklist, notBannedIp would still be true therefore
allowing access to the ip address.

The solution was to make sure if we found an ip and the mode was set to
'deny' that notBannedIp should be set back to false.