Overhauling Error Handling
Instead of trying to decided how to handle errors and add setting after
setting, now you can handle the error however you like. If an IP is
denied, `next()` will be called with a `IpDeniedError` passed as the
argument. You can then check for the error after the middleware and
decided how you want to handle it.